refactor: поля профиля перенесены в константы
