+++ src/lib/constants.ts
@@ -4,3 +4,17 @@
 export const LAPTOP = '(max-width: 1024px)';
 export const TABLET = '(max-width: 768px)';
 export const MOBILE = '(max-width: 550px)';
+
+// Константа для отображаемых названий полей профиля
+export const PROFILE_FIELDS = {
+  tabNum: 'Табельный номер',
+  surname: 'Фамилия',
+  name: 'Имя',
+  patronymic: 'Отчество',
+  pos: 'Должность',
+  department: 'Подразделение',
+  login: 'Логин',
+  email: 'Почта',
+  password: 'Пароль',
+  phoneNum: 'Телефон',
+} as const;
